update latest CCG branch

+++ parliament/vueapp/src/components/Issues.vue
@@ -15,6 +15,109 @@
       </button>
     </div> <!-- /page error -->
 
+    <!-- search & paging -->
+    <div>
+      <!-- page size -->
+      <select number
+        class="form-control page-select"
+        v-model="query.length"
+        @change="updatePaging">
+        <option value="10">
+          10 per page
+        </option>
+        <option value="50">
+          50 per page
+        </option>
+        <option value="100">
+          100 per page
+        </option>
+        <option value="200">
+          200 per page
+        </option>
+        <option value="500">
+          500 per page
+        </option>
+        <option value="1000">
+          1000 per page
+        </option>
+      </select> <!-- /page size -->
+      <!-- paging -->
+      <b-pagination size="sm"
+        v-model="currentPage"
+        :limit="5"
+        hide-ellipsis
+        :total-rows="recordsFiltered"
+        :per-page="parseInt(query.length)"
+        @input="updatePaging">
+      </b-pagination> <!-- paging -->
+      <!-- page info -->
+      <div class="pagination-info"
+        v-b-tooltip.hover>
+        Showing
+        <span v-if="recordsFiltered">
+          {{ start + 1 }}
+        </span>
+        <span v-else>
+          {{ start }}
+        </span>
+        <span v-if="recordsFiltered">
+          - {{ Math.min((start + query.length), recordsFiltered) }}
+        </span>
+        of {{ recordsFiltered }} entries
+      </div>
+      <!-- /page info -->
+      <!-- remove/cancel all issues button -->
+      <button v-if="loggedIn && issues && issues.length"
+        class="btn btn-outline-danger btn-sm cursor-pointer"
+        v-b-tooltip.hover.bottom
+        title="Remove ALL acknowledged issues across the ENTIRE Parliament"
+        @click="removeAllAcknowledgedIssues">
+        <span class="fa fa-trash fa-fw">
+        </span>
+        <transition name="visibility">
+          <span v-if="removeAllAcknowledgedIssuesConfirm">
+            Click to confirm
+          </span>
+        </transition>
+      </button>
+      <transition name="slide-fade">
+        <button class="btn btn-outline-warning btn-sm cursor-pointer"
+          v-if="loggedIn && issues && issues.length && removeAllAcknowledgedIssuesConfirm"
+          v-b-tooltip.hover.bottom
+          title="Cancel removing ALL acknowledged issues"
+          @click="cancelRemoveAllAcknowledgedIssues">
+          <span class="fa fa-ban fa-fw">
+          </span>&nbsp;
+          Cancel
+        </button>
+      </transition>
+      <!-- /remove/cancel all issues button -->
+      <!-- search -->
+      <div class="input-group input-group-sm pull-right issue-search">
+        <div class="input-group-prepend">
+          <span class="input-group-text input-group-text-fw">
+            <span class="fa fa-search fa-fw">
+            </span>
+          </span>
+        </div>
+        <input type="text"
+          class="form-control"
+          v-model="searchTerm"
+          @input="debounceSearchInput"
+          @keyup.enter="debounceSearchInput"
+          placeholder="Begin typing to search for issues"
+        />
+        <span class="input-group-append">
+          <button type="button"
+            @click="clear"
+            class="btn btn-outline-secondary">
+            <span class="fa fa-close">
+            </span>
+          </button>
+        </span>
+      </div> <!-- /search -->
+    </div> <!-- /search & paging -->
+
     <!-- issues table -->
     <table class="table table-hover table-sm">
       <thead>
@@ -128,7 +231,7 @@
               class="fa fa-sort-desc fa-fw">
             </span>
           </th>
-          <th v-if="loggedIn && issues.length"
+          <th v-if="loggedIn && issues && issues.length"
             width="120px"
             scope="col">
             <span v-if="atLeastOneIssueSelected">
@@ -177,15 +280,15 @@
                 <span class="fa fa-check fa-fw">
                 </span>
               </button> <!-- /acknowledge issues button -->
+              <!-- remove selected issues button -->
+              <button class="btn btn-outline-danger btn-xs pull-right cursor-pointer"
+                v-b-tooltip.hover.bottom-right
+                title="Remove selected acknowledged issues"
+                @click="removeSelectedAcknowledgedIssues">
+                <span class="fa fa-trash fa-fw">
+                </span>
+              </button> <!-- /remove selected issues button -->
             </span>
-            <!-- remove all issues button -->
-            <button class="btn btn-outline-primary btn-xs pull-right cursor-pointer"
-              v-b-tooltip.hover.bottom-right
-              title="Remove ALL acknowledged issues"
-              @click="removeAllAcknowledgedIssues">
-              <span class="fa fa-trash fa-fw">
-              </span>
-            </button> <!-- /remove all issues button -->
           </th>
         </tr>
       </thead>
@@ -245,9 +348,16 @@
     <div v-if="!loading && (!issues || !issues.length)"
       class="info-area vertical-center text-center">
       <div class="text-muted mt-5">
-        <span class="fa fa-3x fa-smile-o text-muted-more">
+        <span v-if="!searchTerm">
+          <span class="fa fa-3x fa-smile-o text-muted-more">
+          </span>
+          No issues in your Parliament
         </span>
-        No issues in your Parliament
+        <span v-else>
+          <span class="fa fa-3x fa-folder-open-o text-muted-more">
+          </span>
+          No issues match your search
+        </span>
       </div>
     </div> <!-- no clusters -->
 
@@ -260,6 +370,7 @@ import ParliamentService from './parliament.service';
 import IssueActions from './IssueActions';
 
 let interval;
+let searchInputTimeout;
 
 export default {
   name: 'Issues',
@@ -275,7 +386,15 @@ export default {
       // page data
       issues: [],
       allIssuesSelected: false,
-      atLeastOneIssueSelected: false
+      atLeastOneIssueSelected: false,
+      // pagination (always start on the first page)
+      start: 0,
+      currentPage: 1,
+      recordsFiltered: 0,
+      // searching
+      searchTerm: undefined,
+      // remove ALL ack issues confirm (double click)
+      removeAllAcknowledgedIssuesConfirm: false
     };
   },
   computed: {
@@ -286,12 +405,13 @@ export default {
     refreshInterval: function () {
       return this.$store.state.refreshInterval;
     },
-    // table sorting
+    // table sorting and paging
     query: {
       get: function () {
         return {
           sort: this.$route.query.sort,
-          order: this.$route.query.order || 'desc'
+          order: this.$route.query.order || 'desc',
+          length: this.$route.query.length || 50
         };
       },
       set: function (val) {
@@ -343,7 +463,9 @@ export default {
 
       this.query = {
         sort: sort,
-        order: order
+        order: order,
+        start: this.start,
+        length: this.query.length
       };
     },
     getIssueTrackingId: function (issue) {
@@ -363,14 +485,39 @@ export default {
 
       return '';
     },
+    cancelRemoveAllAcknowledgedIssues: function () {
+      this.removeAllAcknowledgedIssuesConfirm = false;
+    },
     removeAllAcknowledgedIssues: function () {
-      ParliamentService.removeAllAcknowledgedIssues()
+      if (!this.removeAllAcknowledgedIssuesConfirm) {
+        this.removeAllAcknowledgedIssuesConfirm = true;
+        return;
+      }
+
+      if (this.removeAllAcknowledgedIssuesConfirm) {
+        this.removeAllAcknowledgedIssuesConfirm = false;
+        ParliamentService.removeAllAcknowledgedIssues()
+          .then((data) => {
+            this.error = '';
+            this.loadData(); // fetch new issues
+          })
+          .catch((error) => {
+            this.error = error.text || 'Error removing all acknowledged issues.';
+          });
+      }
+    },
+    removeSelectedAcknowledgedIssues: function () {
+      let selectedIssues = this.getSelectedIssues();
+
+      ParliamentService.removeSelectedAcknowledgedIssues(selectedIssues)
         .then((data) => {
+          this.allIssuesSelected = false;
+          this.atLeastOneIssueSelected = false;
           this.error = '';
-          this.issues = data.issues;
+          this.loadData(); // fetch new issues
         })
         .catch((error) => {
-          this.error = error.text || 'Error removing all acknowledged issues.';
+          this.error = error.text || `Unable to remove ${selectedIssues.length} issues`;
         });
     },
     toggleIssue: function (issue) {
@@ -450,16 +597,51 @@ export default {
           this.error = error.text || `Unable to unignore ${selectedIssues.length} issues`;
         });
     },
+    updatePaging: function () {
+      this.start = (this.currentPage - 1) * this.query.length;
+
+      this.query = {
+        sort: this.query.sort,
+        order: this.query.order,
+        length: this.query.length
+      };
+
+      this.loadData();
+    },
+    debounceSearchInput: function () {
+      if (searchInputTimeout) { clearTimeout(searchInputTimeout); }
+      // debounce the input so it only issues a request after keyups cease for 400ms
+      searchInputTimeout = setTimeout(() => {
+        searchInputTimeout = null;
+        this.loadData();
+      }, 400);
+    },
+    clear: function () {
+      this.searchTerm = undefined;
+      this.loadData();
+    },
     /* helper functions ---------------------------------------------------- */
     loadData: function () {
-      let query = {}; // set up query parameters (order and sort)
-      if (this.query.sort) { query = { order: this.query.order, sort: this.query.sort }; }
+      let query = { // set up query parameters (order, sort, paging)
+        start: this.start,
+        length: this.query.length
+      };
+
+      if (this.query.sort) {
+        query.sort = this.query.sort;
+        query.order = this.query.order;
+      }
+
+      if (this.searchTerm) {
+        query.filter = this.searchTerm;
+      }
 
       ParliamentService.getIssues(query)
         .then((data) => {
           this.error = '';
           this.loading = false;
           this.issues = data.issues;
+          this.recordsFiltered = data.recordsFiltered;
         })
         .catch((error) => {
           this.loading = false;
@@ -492,3 +674,61 @@ export default {
   }
 };
 </script>
+
+<style scoped>
+.pagination {
+  display: inline-flex;
+}
+
+select.page-select {
+  width: 120px;
+  font-size: .8rem;
+  display: inline-flex;
+  height: 31px !important;
+  margin-top: 1px;
+  margin-right: -5px;
+  margin-bottom: var(--px-xs);
+  padding-top: var(--px-xs);
+  padding-bottom: var(--px-xs);
+  border-right: none;
+  -webkit-appearance: none;
+}
+
+.pagination-info {
+  display: inline-block;
+  font-size: .8rem;
+  color: #495057;
+  border: 1px solid #CED4DA;
+  padding: 5px 10px;
+  margin-left: -6px;
+  border-radius: 0 var(--px-sm) var(--px-sm) 0;
+  background-color: #FFFFFF;
+}
+
+.input-group.issue-search {
+  flex-wrap: none;
+  width: auto;
+  min-width: 40%;
+}
+
+/* button animation */
+.slide-fade-enter-active {
+  transition: all .5s ease;
+}
+.slide-fade-leave-active {
+  transition: all .5s ease;
+}
+.slide-fade-enter, .slide-fade-leave-to {
+  transform: translateX(-10px);
+  opacity: 0;
+}
+.visibility-enter-active {
+  transition: all .5s ease;
+}
+.visibility-leave-active {
+  transition: all .5s cubic-bezier(1.0, 0.5, 0.8, 1.0);
+}
+.visibility-enter, .visibility-leave-to {
+  opacity: 0;
+}
+</style>
